Feature - is_same_as/is_not_same_as
Added two new methods to the AssertionBuilder class, is_same_as(other_thing) and is_not_same_as(other_thing). These methods do identity comparison (via is or is not), which did not appear to be present in the library. Tests for these methods were added in as tests/test_same_as.py